requirements.txt
Flask
Flask-Cors
duckdb
pandas
langchain_groq
python-dotenv
gunicorn